+++ openBCIGanglionSample.js
@@ -1,0 +1,327 @@
+'use strict';
+const k = require('./openBCIConstants');
+
+module.exports = {
+  convert18bitAsInt32,
+  convert19bitAsInt32,
+  decompressDeltas18Bit,
+  decompressDeltas19Bit,
+  sampleCompressedData: (sampleNumber) => {
+    return new Buffer(
+      [
+        sampleNumber, // 0
+        0b00000000, // 0
+        0b00000000, // 1
+        0b00000000, // 2
+        0b00000000, // 3
+        0b00001000, // 4
+        0b00000000, // 5
+        0b00000101, // 6
+        0b00000000, // 7
+        0b00000000, // 8
+        0b01001000, // 9
+        0b00000000, // 10
+        0b00001001, // 11
+        0b11110000, // 12
+        0b00000001, // 13
+        0b10110000, // 14
+        0b00000000, // 15
+        0b00110000, // 16
+        0b00000000, // 17
+        0b00001000  // 18
+      ]);
+  },
+  sampleImpedanceChannel1: () => {
+    return new Buffer([k.OBCIGanglionByteIdImpedanceChannel1, 0, 0, 1]);
+  },
+  sampleImpedanceChannel2: () => {
+    return new Buffer([k.OBCIGanglionByteIdImpedanceChannel2, 0, 0, 1]);
+  },
+  sampleImpedanceChannel3: () => {
+    return new Buffer([k.OBCIGanglionByteIdImpedanceChannel3, 0, 0, 1]);
+  },
+  sampleImpedanceChannel4: () => {
+    return new Buffer([k.OBCIGanglionByteIdImpedanceChannel4, 0, 0, 1]);
+  },
+  sampleImpedanceChannelReference: () => {
+    return new Buffer([k.OBCIGanglionByteIdImpedanceChannelReference, 0, 0, 1]);
+  },
+  sampleMultiBytePacket: (data) => {
+    const bufPre = new Buffer([k.OBCIGanglionByteIdMultiPacket]);
+    return Buffer.concat([bufPre, data]);
+  },
+  sampleMultiBytePacketStop: (data) => {
+    const bufPre = new Buffer([k.OBCIGanglionByteIdMultiPacketStop]);
+    return Buffer.concat([bufPre, data]);
+  },
+  sampleOtherData: (data) => {
+    const bufPre = new Buffer([255]);
+    return Buffer.concat([bufPre, data]);
+  },
+  sampleUncompressedData: () => {
+    return new Buffer(
+      [
+        0b00000000, // 0
+        0b00000000, // 1
+        0b00000000, // 2
+        0b00000001, // 3
+        0b00000000, // 4
+        0b00000000, // 5
+        0b00000010, // 6
+        0b00000000, // 7
+        0b00000000, // 8
+        0b00000011, // 9
+        0b00000000, // 10
+        0b00000000, // 11
+        0b00000100, // 12
+        0b00000001, // 13
+        0b00000010, // 14
+        0b00000011, // 15
+        0b00000100, // 16
+        0b00000101, // 17
+        0b00000110, // 18
+        0b00000111  // 19
+      ]);
+  }
+};
+
+/**
+ * Converts a two byte buffer into a 32bit number
+ * @param twoByteBuffer {Buffer} - two's complement number
+ * @return {number}
+ *  Converted number
+ */
+function interpret16bitAsInt32 (twoByteBuffer) {
+  var prefix = 0;
+
+  if (twoByteBuffer[0] > 127) {
+    // console.log('\t\tNegative number')
+    prefix = 65535; // 0xFFFF
+  }
+
+  return (prefix << 16) | (twoByteBuffer[0] << 8) | twoByteBuffer[1];
+}
+
+/**
+ * Converts a special ganglion 18 bit compressed number
+ *  The compressions uses the LSB, bit 1, as the signed bit, instead of using
+ *  the MSB. Therefore you must not look to the MSB for a sign extension, one
+ *  must look to the LSB, and the same rules applies, if it's a 1, then it's a
+ *  negative and if it's 0 then it's a positive number.
+ * @param threeByteBuffer {Buffer}
+ *  A 3-byte buffer with only 18 bits of actual data.
+ * @return {number} A signed integer.
+ */
+function convert18bitAsInt32 (threeByteBuffer) {
+  let prefix = 0;
+
+  if (threeByteBuffer[2] & 0x01 > 0) {
+    // console.log('\t\tNegative number')
+    prefix = 0b11111111111111;
+  }
+
+  return (prefix << 18) | (threeByteBuffer[0] << 16) | (threeByteBuffer[1] << 8) | threeByteBuffer[2];
+}
+
+/**
+ * Converts a special ganglion 19 bit compressed number
+ *  The compressions uses the LSB, bit 1, as the signed bit, instead of using
+ *  the MSB. Therefore you must not look to the MSB for a sign extension, one
+ *  must look to the LSB, and the same rules applies, if it's a 1, then it's a
+ *  negative and if it's 0 then it's a positive number.
+ * @param threeByteBuffer {Buffer}
+ *  A 3-byte buffer with only 19 bits of actual data.
+ * @return {number} A signed integer.
+ */
+function convert19bitAsInt32 (threeByteBuffer) {
+  let prefix = 0;
+
+  if (threeByteBuffer[2] & 0x01 > 0) {
+    // console.log('\t\tNegative number')
+    prefix = 0b1111111111111;
+  }
+
+  return (prefix << 19) | (threeByteBuffer[0] << 16) | (threeByteBuffer[1] << 8) | threeByteBuffer[2];
+}
+
+/**
+ * Called to when a compressed packet is received.
+ * @param buffer {Buffer} Just the data portion of the sample. So 18 bytes.
+ * @return {Array} - An array of deltas of shape 2x4 (2 samples per packet
+ *  and 4 channels per sample.)
+ * @private
+ */
+function decompressDeltas18Bit (buffer) {
+  let D = new Array(k.OBCIGanglionSamplesPerPacket); // 2
+  D[0] = [0, 0, 0, 0];
+  D[1] = [0, 0, 0, 0];
+
+  let receivedDeltas = [];
+  for (let i = 0; i < k.OBCIGanglionSamplesPerPacket; i++) {
+    receivedDeltas.push([0, 0, 0, 0]);
+  }
+
+  let miniBuf;
+
+  // Sample 1 - Channel 1
+  miniBuf = new Buffer(
+    [
+      (buffer[0] >> 6),
+      ((buffer[0] & 0x3F) << 2) | (buffer[1] >> 6),
+      ((buffer[1] & 0x3F) << 2) | (buffer[2] >> 6)
+    ]
+  );
+  receivedDeltas[0][0] = convert18bitAsInt32(miniBuf);
+
+  // Sample 1 - Channel 2
+  miniBuf = new Buffer(
+    [
+      (buffer[2] & 0x3F) >> 4,
+      (buffer[2] << 4) | (buffer[3] >> 4),
+      (buffer[3] << 4) | (buffer[4] >> 4)
+    ]);
+  // miniBuf = new Buffer([(buffer[2] & 0x1F), buffer[3], buffer[4] >> 2]);
+  receivedDeltas[0][1] = convert18bitAsInt32(miniBuf);
+
+  // Sample 1 - Channel 3
+  miniBuf = new Buffer(
+    [
+      (buffer[4] & 0x0F) >> 2,
+      (buffer[4] << 6) | (buffer[5] >> 2),
+      (buffer[5] << 6) | (buffer[6] >> 2)
+    ]);
+  receivedDeltas[0][2] = convert18bitAsInt32(miniBuf);
+
+  // Sample 1 - Channel 4
+  miniBuf = new Buffer(
+    [
+      (buffer[6] & 0x03),
+      buffer[7],
+      buffer[8]
+    ]);
+  receivedDeltas[0][3] = convert18bitAsInt32(miniBuf);
+
+  // Sample 2 - Channel 1
+  miniBuf = new Buffer(
+    [
+      (buffer[9] >> 6),
+      ((buffer[9] & 0x3F) << 2) | (buffer[10] >> 6),
+      ((buffer[10] & 0x3F) << 2) | (buffer[11] >> 6)
+    ]);
+  receivedDeltas[1][0] = convert18bitAsInt32(miniBuf);
+
+  // Sample 2 - Channel 2
+  miniBuf = new Buffer(
+    [
+      (buffer[11] & 0x3F) >> 4,
+      (buffer[11] << 4) | (buffer[12] >> 4),
+      (buffer[12] << 4) | (buffer[13] >> 4)
+    ]);
+  receivedDeltas[1][1] = convert18bitAsInt32(miniBuf);
+
+  // Sample 2 - Channel 3
+  miniBuf = new Buffer(
+    [
+      (buffer[13] & 0x0F) >> 2,
+      (buffer[13] << 6) | (buffer[14] >> 2),
+      (buffer[14] << 6) | (buffer[15] >> 2)
+    ]);
+  receivedDeltas[1][2] = convert18bitAsInt32(miniBuf);
+
+  // Sample 2 - Channel 4
+  miniBuf = new Buffer([(buffer[15] & 0x03), buffer[16], buffer[17]]);
+  receivedDeltas[1][3] = convert18bitAsInt32(miniBuf);
+
+  return receivedDeltas;
+}
+
+/**
+ * Called to when a compressed packet is received.
+ * @param buffer {Buffer} Just the data portion of the sample. So 19 bytes.
+ * @return {Array} - An array of deltas of shape 2x4 (2 samples per packet
+ *  and 4 channels per sample.)
+ * @private
+ */
+function decompressDeltas19Bit (buffer) {
+  let D = new Array(k.OBCIGanglionSamplesPerPacket); // 2
+  D[0] = [0, 0, 0, 0];
+  D[1] = [0, 0, 0, 0];
+
+  let receivedDeltas = [];
+  for (let i = 0; i < k.OBCIGanglionSamplesPerPacket; i++) {
+    receivedDeltas.push([0, 0, 0, 0]);
+  }
+
+  let miniBuf;
+
+  // Sample 1 - Channel 1
+  miniBuf = new Buffer(
+    [
+      (buffer[0] >> 5),
+      ((buffer[0] & 0x1F) << 3) | (buffer[1] >> 5),
+      ((buffer[1] & 0x1F) << 3) | (buffer[2] >> 5)
+    ]
+  );
+  receivedDeltas[0][0] = convert19bitAsInt32(miniBuf);
+
+  // Sample 1 - Channel 2
+  miniBuf = new Buffer(
+    [
+      (buffer[2] & 0x1F) >> 2,
+      (buffer[2] << 6) | (buffer[3] >> 2),
+      (buffer[3] << 6) | (buffer[4] >> 2)
+    ]);
+  // miniBuf = new Buffer([(buffer[2] & 0x1F), buffer[3], buffer[4] >> 2]);
+  receivedDeltas[0][1] = convert19bitAsInt32(miniBuf);
+
+  // Sample 1 - Channel 3
+  miniBuf = new Buffer(
+    [
+      ((buffer[4] & 0x03) << 1) | (buffer[5] >> 7),
+      ((buffer[5] & 0x7F) << 1) | (buffer[6] >> 7),
+      ((buffer[6] & 0x7F) << 1) | (buffer[7] >> 7)
+    ]);
+  receivedDeltas[0][2] = convert19bitAsInt32(miniBuf);
+
+  // Sample 1 - Channel 4
+  miniBuf = new Buffer(
+    [
+      ((buffer[7] & 0x7F) >> 4),
+      ((buffer[7] & 0x0F) << 4) | (buffer[8] >> 4),
+      ((buffer[8] & 0x0F) << 4) | (buffer[9] >> 4)
+    ]);
+  receivedDeltas[0][3] = convert19bitAsInt32(miniBuf);
+
+  // Sample 2 - Channel 1
+  miniBuf = new Buffer(
+    [
+      ((buffer[9] & 0x0F) >> 1),
+      (buffer[9] << 7) | (buffer[10] >> 1),
+      (buffer[10] << 7) | (buffer[11] >> 1)
+    ]);
+  receivedDeltas[1][0] = convert19bitAsInt32(miniBuf);
+
+  // Sample 2 - Channel 2
+  miniBuf = new Buffer(
+    [
+      ((buffer[11] & 0x01) << 2) | (buffer[12] >> 6),
+      (buffer[12] << 2) | (buffer[13] >> 6),
+      (buffer[13] << 2) | (buffer[14] >> 6)
+    ]);
+  receivedDeltas[1][1] = convert19bitAsInt32(miniBuf);
+
+  // Sample 2 - Channel 3
+  miniBuf = new Buffer(
+    [
+      ((buffer[14] & 0x38) >> 3),
+      ((buffer[14] & 0x07) << 5) | ((buffer[15] & 0xF8) >> 3),
+      ((buffer[15] & 0x07) << 5) | ((buffer[16] & 0xF8) >> 3)
+    ]);
+  receivedDeltas[1][2] = convert19bitAsInt32(miniBuf);
+
+  // Sample 2 - Channel 4
+  miniBuf = new Buffer([(buffer[16] & 0x07), buffer[17], buffer[18]]);
+  receivedDeltas[1][3] = convert19bitAsInt32(miniBuf);
+
+  return receivedDeltas;
+};