+++ changelog.md
@@ -1,3 +1,25 @@
+# 0.3.2
+
+### Bug Fixes
+* Fix bug where disconnect did not set private property `_connected` to false.
+
+# 0.3.1
+
+### Bug Fixes
+* Fix bug with connect
+
+# 0.3.0
+
+### New Features
+* Get accelerometer data from the ganglion! (Previous did not work while streaming)
+
+### Breaking Changes
+* Major change in how bytes are parse based on byte ID.
+
+### Enhancements
+* Refactor file names for clarity
+* Removed dependency `underscore`
+
 # 0.2.0
 
 ### Enhancements